(PC-7565) : Move legacy favorite repository tests
Merge the legacy favorite repository tests into
`tests.core.users.test_repository`.
